// WEB/FootageAngular/src/app/model/interfaces/usuario.ts
// Generated by https://quicktype.io

export interface UsuarioResponse {
    id:           number;
    nombre:       string;
    apellidos:    string;
    username:     string;
    email:        string;
    avatar:       string;
    localizacion: string;
    premium:      boolean;
    rol:          string;
    articulos:    Articulo[];
    meGustas:     Articulo[];
}

export interface Articulo {
    id:       number;
    titulo:   string;
    precio:   number;
    original: boolean;
    foto:    string;
    marca:    Marca;
    talla:    Talla;
}

export interface Marca {
    id:     number;
    nombre: string;
    imagen: string;
}

export interface Talla {
    id:        number;
    nombre:    string;
}